requirements.txt
# API Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database Clients
asyncpg==0.29.0
clickhouse-driver==0.2.6
questdb==1.1.0
psycopg2-binary==2.9.9

# Message Queue
nats-py==2.6.0
asyncio-nats-streaming==0.4.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-prometheus==0.50b0

# Utils
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.26.0
aiohttp==3.9.1
aiodns==3.1.1
charset-normalizer==3.4.1
orjson==3.9.10
ujson==5.8.0
email-validator==2.1.0.post1
pandas==2.1.3
numpy==1.26.2
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0
pylint==3.0.2

# Dagster packages
dagster==1.5.13
dagster-graphql==1.5.13
dagster-webserver==1.5.13
dagster-postgres==0.21.13
dagster-celery==0.21.13
dagster-celery-docker==0.21.13 